feat: migrate mapping to work reliably across platforms
I had a key-rollover issue which was near impossible to solve with the
complicated state-manipulating way that I had been doing things.
Now modifier and key-state are double stored, what the user is pressing,
and what the kbd logically things should be pressed.
That means that synthetic keys and modifiers are separated from what the
user wants to do. That means that popping/pushing keys/modifiers doesn't
interfere with what the user actually wants, creating fewer bugs on
custom macros/sequences.
Additionally, key-state is restored to user state before handling a new
press, temp modifiers are only reset if the state isn't stale, and
each change to the keyboard creates its own entry which is sent to the
OS.
macos specifically cannot handle many things happening on the same
report properly, and as far as latency goes it's one report per ms,
the longest chains are ~5 actions which means that the OS will have
picked up the entire action chain in about 5ms in the worst case.

This reverts commit 0105b98191eb2042317e0dea0ba6c9963d2aeeaa.
