+++ rp2040-kbd/src/keymap.rs
@@ -1,3 +1,5 @@
+mod mapping;
+
 #[cfg(feature = "serial")]
 use core::fmt::Write;
 use core::hint::unreachable_unchecked;
@@ -10,7 +12,7 @@ use usbd_hid::descriptor::KeyboardReport;
 
 use crate::keyboard::debounce::PinDebouncer;
 use crate::keyboard::left::LeftButtons;
-use crate::runtime::shared::cores_left::{push_layer_change, push_reboot_and_halt, Producer};
+use crate::runtime::shared::cores_left::{push_layer_change, Producer};
 use rp2040_kbd_lib::keycodes::{KeyCode, Modifier};
 use rp2040_kbd_lib::matrix::{MatrixChange, MatrixUpdate};
 
@@ -45,9 +47,10 @@ fn copy_report(keyboard_report: &KeyboardReport) -> KeyboardReport {
 }
 
 pub struct KeyboardReportState {
+    generation: usize,
     inner_report: KeyboardReport,
-    inner_report_has_change: bool,
-    temp_mods: Option<Modifier>,
+    user_mods: Modifier,
+    user_key_state: [u8; 6],
     outbound_reports: Queue<KeyboardReport, 16>,
     active_layer: KeymapLayer,
     last_perm_layer: Option<KeymapLayer>,
@@ -69,9 +72,10 @@ struct JankState {
 impl KeyboardReportState {
     pub fn new() -> Self {
         Self {
+            generation: 0,
             inner_report: EMPTY_REPORT,
-            inner_report_has_change: true,
-            temp_mods: None,
+            user_mods: Modifier(0),
+            user_key_state: [0; 6],
             outbound_reports: Queue::new(),
             active_layer: KeymapLayer::DvorakSe,
             last_perm_layer: None,
@@ -88,26 +92,60 @@ impl KeyboardReportState {
         }
     }
 
+    pub fn increment_generation(&mut self) {
+        self.generation = self.generation.wrapping_add(1);
+    }
+
     #[cfg(feature = "hiddev")]
     pub fn report(&self) -> Option<&KeyboardReport> {
-        self.outbound_reports
-            .peek()
-            .or_else(|| self.inner_report_has_change.then_some(&self.inner_report))
+        self.outbound_reports.peek()
     }
 
     #[cfg(feature = "hiddev")]
     pub fn accept(&mut self) {
-        if self.outbound_reports.pop_front().is_none() {
-            self.inner_report_has_change = false;
+        self.outbound_reports.pop_front();
+    }
+
+    fn restore_to_user_if_not_stale(&mut self, generation: usize) {
+        if self.generation == generation.wrapping_add(1) {
+            self.restore_to_user_state();
         }
     }
 
-    pub fn push_key(&mut self, key_code: KeyCode) {
-        self.pop_temp_modifiers();
-        for val in &mut self.inner_report.keycodes {
+    fn restore_to_user_state(&mut self) {
+        self.restore_to_user_mods();
+        self.restore_to_user_keys();
+    }
+
+    fn restore_to_user_mods(&mut self) {
+        if self.inner_report.modifier != self.user_mods.0 {
+            self.inner_report.modifier = self.user_mods.0;
+            self.report_current();
+        }
+    }
+
+    fn restore_to_user_keys(&mut self) {
+        if self.inner_report.keycodes != self.user_key_state {
+            self.inner_report.keycodes = self.user_key_state;
+            self.report_current();
+        }
+    }
+
+    /// Restores state to user, then applies the new key
+    fn push_key(&mut self, key_code: KeyCode) {
+        self.push_key_raw(key_code);
+        self.report_current();
+    }
+
+    fn push_key_raw(&mut self, key_code: KeyCode) {
+        Self::push_key_to_arr(key_code, &mut self.user_key_state);
+        Self::push_key_to_arr(key_code, &mut self.inner_report.keycodes);
+    }
+
+    fn push_key_to_arr(key_code: KeyCode, arr: &mut [u8; 6]) {
+        for val in &mut *arr {
             if *val == 0 || *val == key_code.0 {
                 *val = key_code.0;
-                self.inner_report_has_change = true;
                 return;
             }
         }
@@ -116,27 +154,78 @@ impl KeyboardReportState {
             #[cfg(feature = "serial")]
             {
                 let _ = crate::runtime::shared::usb::acquire_usb()
-                    .write_fmt(format_args!("Pre: {:?}\r\n", self.inner_report.keycodes));
+                    .write_fmt(format_args!("Pre: {arr:?}\r\n"));
             }
-            copy_within_unchecked(&mut self.inner_report.keycodes, 1, 5, 0);
-            *self.inner_report.keycodes.get_unchecked_mut(5) = key_code.0;
+            copy_within_unchecked(arr, 1, 5, 0);
+            *arr.get_unchecked_mut(5) = key_code.0;
             #[cfg(feature = "serial")]
             {
                 let _ = crate::runtime::shared::usb::acquire_usb()
-                    .write_fmt(format_args!("Post: {:?}\r\n", self.inner_report.keycodes));
+                    .write_fmt(format_args!("Post: {arr:?}\r\n"));
             }
         }
-        self.inner_report_has_change = true;
     }
 
-    pub fn pop_key(&mut self, key_code: KeyCode) {
+    fn report_current(&mut self) {
+        self.outbound_reports
+            .push_back(copy_report(&self.inner_report));
+    }
+
+    fn pop_key(&mut self, key_code: KeyCode) {
+        if Self::pop_key_from_arr(key_code, &mut self.user_key_state) {
+            self.inner_report.keycodes = self.user_key_state;
+            self.report_current();
+        }
+    }
+
+    fn temp_modify(
+        &mut self,
+        key_code: KeyCode,
+        add_modifiers: &[Modifier],
+        remove_modifiers: &[Modifier],
+    ) {
+        self.push_temp_modifiers(add_modifiers);
+        self.temp_remove_modifiers(remove_modifiers);
+        self.push_temp_key(key_code);
+    }
+
+    fn push_temp_modifiers(&mut self, modifier: &[Modifier]) {
+        for m in modifier {
+            if self.inner_report.modifier & m.0 == 0 {
+                self.inner_report.modifier |= m.0;
+                self.report_current();
+            }
+        }
+    }
+
+    fn temp_remove_modifiers(&mut self, modifier: &[Modifier]) {
+        for m in modifier {
+            if self.inner_report.modifier & m.0 != 0 {
+                self.inner_report.modifier &= !m.0;
+                self.report_current();
+            }
+        }
+    }
+
+    fn push_temp_key(&mut self, key_code: KeyCode) {
+        Self::push_key_to_arr(key_code, &mut self.inner_report.keycodes);
+        self.report_current();
+    }
+
+    fn pop_temp_key(&mut self, key_code: KeyCode) {
+        if Self::pop_key_from_arr(key_code, &mut self.inner_report.keycodes) {
+            self.report_current();
+        }
+    }
+
+    pub fn pop_key_from_arr(key_code: KeyCode, arr: &mut [u8; 6]) -> bool {
         let mut at_ind = None;
-        for (ind, val) in &mut self.inner_report.keycodes.iter().enumerate() {
+        for (ind, val) in arr.iter().enumerate() {
             if *val == key_code.0 {
                 at_ind = Some(ind);
                 break;
             } else if *val == 0 {
-                return;
+                return false;
             }
         }
         if let Some(ind) = at_ind {
@@ -144,100 +233,76 @@ impl KeyboardReportState {
                 #[cfg(feature = "serial")]
                 {
                     let _ = crate::runtime::shared::usb::acquire_usb()
-                        .write_fmt(format_args!("Pre: {:?}\r\n", self.inner_report.keycodes));
+                        .write_fmt(format_args!("Pre: {arr:?}\r\n"));
                 }
-                self.pop_copy_back(ind);
+                Self::pop_copy_back_arr(ind, arr);
                 #[cfg(feature = "serial")]
                 {
                     let _ = crate::runtime::shared::usb::acquire_usb()
-                        .write_fmt(format_args!("Post: {:?}\r\n", self.inner_report.keycodes));
+                        .write_fmt(format_args!("Post: {arr:?}\r\n"));
                 }
             }
-            self.outbound_reports
-                .push_back(copy_report(&self.inner_report));
+            true
+        } else {
+            false
         }
     }
 
-    unsafe fn pop_copy_back(&mut self, ind: usize) {
-        *self.inner_report.keycodes.get_unchecked_mut(ind) = 0;
+    unsafe fn pop_copy_back_arr(ind: usize, arr: &mut [u8; 6]) {
+        *arr.get_unchecked_mut(ind) = 0;
         match ind {
             0 => {
-                copy_within_unchecked(&mut self.inner_report.keycodes, 1, 5, 0);
+                copy_within_unchecked(arr, 1, 5, 0);
                 // Keys are shifted back by one, need to clear last or there'll be a duplication
-                *self.inner_report.keycodes.get_unchecked_mut(5) = 0;
+                *arr.get_unchecked_mut(5) = 0;
             }
             1 => {
-                copy_within_unchecked(&mut self.inner_report.keycodes, 2, 4, 1);
-                *self.inner_report.keycodes.get_unchecked_mut(5) = 0;
+                copy_within_unchecked(arr, 2, 4, 1);
+                *arr.get_unchecked_mut(5) = 0;
             }
             2 => {
-                copy_within_unchecked(&mut self.inner_report.keycodes, 3, 3, 2);
-                *self.inner_report.keycodes.get_unchecked_mut(5) = 0;
+                copy_within_unchecked(arr, 3, 3, 2);
+                *arr.get_unchecked_mut(5) = 0;
             }
             3 => {
-                copy_within_unchecked(&mut self.inner_report.keycodes, 4, 2, 3);
-                *self.inner_report.keycodes.get_unchecked_mut(5) = 0;
+                copy_within_unchecked(arr, 4, 2, 3);
+                *arr.get_unchecked_mut(5) = 0;
             }
             4 => {
-                let old = *self.inner_report.keycodes.get_unchecked(5);
-                *self.inner_report.keycodes.get_unchecked_mut(4) = old;
-                *self.inner_report.keycodes.get_unchecked_mut(5) = 0;
+                let old = *arr.get_unchecked(5);
+                *arr.get_unchecked_mut(4) = old;
+                *arr.get_unchecked_mut(5) = 0;
             }
             5 => {}
             _ => unreachable_unchecked(),
         }
     }
 
-    fn temp_modify(&mut self, key_code: KeyCode, add_mods: &[Modifier], pop_mods: &[Modifier]) {
-        self.temp_mods = Some(Modifier(self.inner_report.modifier));
-        let mut new_mods = self.inner_report.modifier;
-        for md in add_mods {
-            new_mods |= md.0;
-        }
-        for md in pop_mods {
-            new_mods &= !md.0;
-        }
-        if new_mods != self.inner_report.modifier {
-            self.inner_report.modifier = new_mods;
-            // Macos can't handle a change in modifier state at the same time as a key press,
-            // send the modifier change first, then the full report
-            self.outbound_reports
-                .push_back(copy_report(&self.inner_report));
-            self.inner_report_has_change = true;
-        }
-        self.push_key(key_code);
-        self.inner_report_has_change = true;
-        self.outbound_reports
-            .push_back(copy_report(&self.inner_report));
-    }
-
     #[inline]
     fn push_modifier(&mut self, modifier: Modifier) {
-        self.pop_temp_modifiers();
-        self.inner_report.modifier |= modifier.0;
-        self.inner_report_has_change = true;
+        if self.user_mods.0 & modifier.0 == 0 {
+            self.user_mods.0 |= modifier.0;
+        }
+        if self.inner_report.modifier & modifier.0 == 0 {
+            self.inner_report.modifier |= modifier.0;
+            self.report_current();
+        }
     }
 
     #[inline]
     fn pop_modifier(&mut self, modifier: Modifier) {
-        self.pop_temp_modifiers();
-        self.inner_report.modifier &= !modifier.0;
-        self.inner_report_has_change = true;
-    }
-
-    #[inline]
-    fn pop_temp_modifiers(&mut self) {
-        if let Some(temp) = self.temp_mods.take() {
-            self.inner_report.modifier = temp.0;
+        if self.user_mods.0 & modifier.0 != 0 {
+            self.user_mods.0 &= !modifier.0;
+            if self.inner_report.modifier != self.user_mods.0 {
+                self.inner_report.modifier = self.user_mods.0;
+                self.report_current();
+            }
         }
     }
 
     #[inline]
-    fn has_modifier(&self, modifier: Modifier) -> bool {
-        self.temp_mods.map_or_else(
-            || self.inner_report.modifier & modifier.0 != 0,
-            |tm| tm.0 & modifier.0 != 0,
-        )
+    fn has_user_modifier(&self, modifier: Modifier) -> bool {
+        self.user_mods.0 & modifier.0 != 0
     }
 
     /// Reset report on all layer switches
@@ -294,6 +359,7 @@ pub trait KeyboardButton {
 
 #[derive(Copy, Clone, Debug)]
 pub struct LastPressState {
+    generation: usize,
     layer: KeymapLayer,
     last_perm_layer: Option<KeymapLayer>,
 }
@@ -312,8 +378,9 @@ impl PinStructState {
     // Keys are individually very rarely pressed compared to the scan-loop latency,
     // this pretty small function not being inlined makes quite the difference.
     #[inline(never)]
-    fn update_last_state(&mut self, current_state: &mut KeyboardReportState) {
+    fn update_last_state(&mut self, current_state: &KeyboardReportState) {
         self.last_state = Some(LastPressState {
+            generation: current_state.generation,
             layer: current_state.active_layer,
             last_perm_layer: current_state.last_perm_layer,
         });
@@ -344,8 +411,10 @@ macro_rules! impl_check_update {
                     if let Some(prev) = self.0.last_state.take() {
                         self.on_release(prev, keyboard_report_state, producer);
                     } else {
+                        keyboard_report_state.restore_to_user_state();
                         self.on_press(keyboard_report_state, producer);
                         self.0.update_last_state(keyboard_report_state);
+                        keyboard_report_state.increment_generation();
                     }
                     return true;
                 }
@@ -525,8 +594,10 @@ macro_rules! handle_update_right {
             if let Some(prev) = $field.0.last_state.take() {
                 $field.on_release(prev, $state, $producer);
             } else {
+                $state.restore_to_user_state();
                 $field.on_press($state, $producer);
                 $field.0.update_last_state($state);
+                $state.increment_generation();
             }
         }
     }};
@@ -867,2808 +938,16 @@ fn rotate_layer(
     }
 }
 
+#[macro_export]
 macro_rules! temp_layer {
     ($layer: pat) => {
         (_, $layer)
     };
 }
 
+#[macro_export]
 macro_rules! base_layer {
     ($layer: pat) => {
         (Some($layer), _) | (None, $layer)
     };
-}
-
-impl KeyboardButton for LeftRow0Col0 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, _producer: &Producer) {
-        keyboard_report_state.push_key(KeyCode::TAB);
-    }
-    fn on_release(
-        &mut self,
-        _last_press_state: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-        keyboard_report_state.pop_key(KeyCode::TAB);
-    }
-}
-
-impl KeyboardButton for LeftRow0Col1 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, _producer: &Producer) {
-        match (
-            keyboard_report_state.last_perm_layer,
-            keyboard_report_state.active_layer,
-        ) {
-            temp_layer!(KeymapLayer::Raise) => {
-                keyboard_report_state.push_key(KeyCode::F1);
-            }
-            temp_layer!(KeymapLayer::Lower | KeymapLayer::LowerAnsi | KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.temp_modify(KeyCode::N1, &[Modifier::LEFT_SHIFT], &[]);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.push_key(KeyCode::COMMA);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.push_key(KeyCode::N1);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                if keyboard_report_state.has_modifier(Modifier::ANY_SHIFT) {
-                    // Shifted, `SHIFT + 2` -> "
-                    keyboard_report_state.jank.pressing_double_quote = true;
-                    keyboard_report_state.push_key(KeyCode::N2);
-                } else {
-                    // Not shifted, \ -> '
-                    keyboard_report_state.jank.pressing_single_quote = true;
-                    keyboard_report_state.push_key(KeyCode::BACKSLASH);
-                }
-            }
-            _ => {}
-        }
-    }
-
-    fn on_release(
-        &mut self,
-        last_press_state: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-        match (last_press_state.last_perm_layer, last_press_state.layer) {
-            temp_layer!(KeymapLayer::Raise) => {
-                keyboard_report_state.pop_key(KeyCode::F1);
-            }
-            temp_layer!(KeymapLayer::Lower | KeymapLayer::LowerAnsi | KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.pop_temp_modifiers();
-                keyboard_report_state.pop_key(KeyCode::N1);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.pop_key(KeyCode::COMMA);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                if keyboard_report_state.jank.pressing_double_quote {
-                    keyboard_report_state.pop_key(KeyCode::N2);
-                    keyboard_report_state.jank.pressing_double_quote = false;
-                }
-                if keyboard_report_state.jank.pressing_single_quote {
-                    keyboard_report_state.jank.pressing_single_quote = false;
-                    keyboard_report_state.pop_key(KeyCode::BACKSLASH);
-                }
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.pop_key(KeyCode::N1);
-            }
-
-            _ => {}
-        }
-    }
-}
-
-impl KeyboardButton for LeftRow0Col2 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, _producer: &Producer) {
-        match (
-            keyboard_report_state.last_perm_layer,
-            keyboard_report_state.active_layer,
-        ) {
-            temp_layer!(KeymapLayer::Settings) => {}
-            temp_layer!(KeymapLayer::Num) => {}
-            temp_layer!(KeymapLayer::Raise) => {
-                keyboard_report_state.push_key(KeyCode::F2);
-            }
-            temp_layer!(KeymapLayer::LowerAnsi) => {
-                keyboard_report_state.temp_modify(KeyCode::N2, &[Modifier::LEFT_SHIFT], &[]);
-            }
-            temp_layer!(KeymapLayer::Lower) => {
-                keyboard_report_state.temp_modify(KeyCode::N2, &[Modifier::RIGHT_ALT], &[]);
-            }
-            // Not particularly different, but it's safest to stick with L_ALT for macos
-            temp_layer!(KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.temp_modify(KeyCode::N2, &[Modifier::LEFT_ALT], &[]);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.push_key(KeyCode::COMMA);
-            }
-            base_layer!(KeymapLayer::DvorakSe) => {
-                if keyboard_report_state.has_modifier(Modifier::LEFT_SHIFT) {
-                    // Need to remove shift for this key to go out, not putting it
-                    // back after though for reasons that I don't remember and may be a bug
-                    keyboard_report_state.temp_modify(
-                        KeyCode::NON_US_BACKSLASH,
-                        &[],
-                        &[Modifier::LEFT_SHIFT],
-                    );
-                    keyboard_report_state.jank.pressing_left_bracket = true;
-                } else {
-                    keyboard_report_state.push_key(KeyCode::COMMA);
-                    keyboard_report_state.jank.pressing_comma = true;
-                }
-            }
-            base_layer!(KeymapLayer::DvorakSeMac) => {
-                if keyboard_report_state.has_modifier(Modifier::LEFT_SHIFT) {
-                    // Need to remove shift for this key to go out, not putting it
-                    // back after though for reasons that I don't remember and may be a bug
-                    keyboard_report_state.temp_modify(KeyCode::GRAVE, &[], &[Modifier::LEFT_SHIFT]);
-                    keyboard_report_state.jank.pressing_left_bracket = true;
-                } else {
-                    keyboard_report_state.push_key(KeyCode::COMMA);
-                    keyboard_report_state.jank.pressing_comma = true;
-                }
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.push_key(KeyCode::N2);
-            }
-            _ => {}
-        }
-    }
-
-    fn on_release(
-        &mut self,
-        prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-        match (prev.last_perm_layer, prev.layer) {
-            temp_layer!(KeymapLayer::Raise) => {
-                keyboard_report_state.pop_key(KeyCode::F2);
-            }
-            temp_layer!(KeymapLayer::LowerAnsi) => {
-                keyboard_report_state.pop_temp_modifiers();
-                keyboard_report_state.pop_key(KeyCode::N2);
-            }
-            temp_layer!(KeymapLayer::Lower | KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.pop_temp_modifiers();
-                keyboard_report_state.pop_key(KeyCode::N2);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.pop_key(KeyCode::COMMA);
-            }
-            base_layer!(KeymapLayer::DvorakSe) => {
-                if keyboard_report_state.jank.pressing_left_bracket {
-                    // These are on the same button and interfere with each other
-                    if !keyboard_report_state.jank.pressing_right_bracket {
-                        keyboard_report_state.pop_temp_modifiers();
-                        keyboard_report_state.pop_key(KeyCode::NON_US_BACKSLASH);
-                    }
-                    keyboard_report_state.jank.pressing_left_bracket = false;
-                }
-                if keyboard_report_state.jank.pressing_comma {
-                    keyboard_report_state.pop_key(KeyCode::COMMA);
-                    keyboard_report_state.jank.pressing_comma = false;
-                }
-            }
-            base_layer!(KeymapLayer::DvorakSeMac) => {
-                if keyboard_report_state.jank.pressing_left_bracket {
-                    // These are on the same button and interfere with each other
-                    if !keyboard_report_state.jank.pressing_right_bracket {
-                        keyboard_report_state.pop_temp_modifiers();
-                        keyboard_report_state.pop_key(KeyCode::GRAVE);
-                    }
-                    keyboard_report_state.jank.pressing_left_bracket = false;
-                }
-                if keyboard_report_state.jank.pressing_comma {
-                    keyboard_report_state.pop_key(KeyCode::COMMA);
-                    keyboard_report_state.jank.pressing_comma = false;
-                }
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.pop_key(KeyCode::N2);
-            }
-
-            _ => {}
-        }
-    }
-}
-
-impl KeyboardButton for LeftRow0Col3 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, _producer: &Producer) {
-        match (
-            keyboard_report_state.last_perm_layer,
-            keyboard_report_state.active_layer,
-        ) {
-            temp_layer!(KeymapLayer::Raise) => {
-                keyboard_report_state.push_key(KeyCode::F3);
-            }
-            temp_layer!(KeymapLayer::Lower | KeymapLayer::LowerAnsi | KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.temp_modify(KeyCode::N3, &[Modifier::LEFT_SHIFT], &[]);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.push_key(KeyCode::DOT);
-            }
-            base_layer!(KeymapLayer::DvorakSe) => {
-                // Button is > or . with and without shift, respectively
-                if keyboard_report_state.has_modifier(Modifier::LEFT_SHIFT) {
-                    // Needs a shift, but that's already pressed
-                    if keyboard_report_state.jank.pressing_left_bracket {
-                        keyboard_report_state.pop_key(KeyCode::NON_US_BACKSLASH);
-                    }
-                    keyboard_report_state.push_key(KeyCode::NON_US_BACKSLASH);
-                    keyboard_report_state.jank.pressing_right_bracket = true;
-                } else {
-                    keyboard_report_state.push_key(KeyCode::DOT);
-                    keyboard_report_state.jank.pressing_dot = true;
-                }
-            }
-            base_layer!(KeymapLayer::DvorakSeMac) => {
-                // Button is > or . with and without shift, respectively
-                if keyboard_report_state.has_modifier(Modifier::LEFT_SHIFT) {
-                    // Needs a shift, but that's already pressed
-                    if keyboard_report_state.jank.pressing_left_bracket {
-                        keyboard_report_state.pop_key(KeyCode::GRAVE);
-                    }
-                    keyboard_report_state.push_key(KeyCode::GRAVE);
-                    keyboard_report_state.jank.pressing_right_bracket = true;
-                } else {
-                    keyboard_report_state.push_key(KeyCode::DOT);
-                    keyboard_report_state.jank.pressing_dot = true;
-                }
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.push_key(KeyCode::N3);
-            }
-
-            _ => {}
-        }
-    }
-
-    fn on_release(
-        &mut self,
-        prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-        match (prev.last_perm_layer, prev.layer) {
-            temp_layer!(KeymapLayer::Raise) => {
-                keyboard_report_state.pop_key(KeyCode::F3);
-            }
-            temp_layer!(KeymapLayer::Lower | KeymapLayer::LowerAnsi | KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.pop_temp_modifiers();
-                keyboard_report_state.pop_key(KeyCode::N3);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.pop_key(KeyCode::DOT);
-            }
-            base_layer!(KeymapLayer::DvorakSe) => {
-                if keyboard_report_state.jank.pressing_right_bracket {
-                    keyboard_report_state.pop_key(KeyCode::NON_US_BACKSLASH);
-                    keyboard_report_state.jank.pressing_right_bracket = false;
-                }
-                if keyboard_report_state.jank.pressing_dot {
-                    keyboard_report_state.pop_key(KeyCode::DOT);
-                    keyboard_report_state.jank.pressing_dot = false;
-                }
-            }
-            base_layer!(KeymapLayer::DvorakSeMac) => {
-                if keyboard_report_state.jank.pressing_right_bracket {
-                    keyboard_report_state.pop_key(KeyCode::GRAVE);
-                    keyboard_report_state.jank.pressing_right_bracket = false;
-                }
-                if keyboard_report_state.jank.pressing_dot {
-                    keyboard_report_state.pop_key(KeyCode::DOT);
-                    keyboard_report_state.jank.pressing_dot = false;
-                }
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.pop_key(KeyCode::N3);
-            }
-
-            _ => {}
-        }
-    }
-}
-
-impl KeyboardButton for LeftRow0Col4 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, _producer: &Producer) {
-        match (
-            keyboard_report_state.last_perm_layer,
-            keyboard_report_state.active_layer,
-        ) {
-            temp_layer!(KeymapLayer::Raise) => {
-                keyboard_report_state.push_key(KeyCode::F4);
-            }
-            temp_layer!(KeymapLayer::LowerAnsi) => {
-                keyboard_report_state.temp_modify(KeyCode::N4, &[Modifier::LEFT_SHIFT], &[]);
-            }
-            temp_layer!(KeymapLayer::Lower | KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.temp_modify(KeyCode::N4, &[Modifier::RIGHT_ALT], &[]);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.push_key(KeyCode::P);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.push_key(KeyCode::P);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.push_key(KeyCode::N4);
-            }
-
-            _ => {}
-        }
-    }
-
-    fn on_release(
-        &mut self,
-        prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-        match (prev.last_perm_layer, prev.layer) {
-            temp_layer!(KeymapLayer::Raise) => {
-                keyboard_report_state.pop_key(KeyCode::F4);
-            }
-            temp_layer!(KeymapLayer::LowerAnsi) => {
-                keyboard_report_state.pop_temp_modifiers();
-                keyboard_report_state.pop_key(KeyCode::N4);
-            }
-            temp_layer!(KeymapLayer::Lower | KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.pop_temp_modifiers();
-                keyboard_report_state.pop_key(KeyCode::N4);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.pop_key(KeyCode::P);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.pop_key(KeyCode::P);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.pop_key(KeyCode::N4);
-            }
-
-            _ => {}
-        }
-    }
-}
-impl KeyboardButton for LeftRow0Col5 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, _producer: &Producer) {
-        match (
-            keyboard_report_state.last_perm_layer,
-            keyboard_report_state.active_layer,
-        ) {
-            temp_layer!(KeymapLayer::Raise) => {
-                keyboard_report_state.push_key(KeyCode::F4);
-            }
-            temp_layer!(KeymapLayer::LowerAnsi) => {
-                keyboard_report_state.temp_modify(KeyCode::N5, &[Modifier::LEFT_SHIFT], &[]);
-            }
-            temp_layer!(KeymapLayer::Lower | KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.temp_modify(KeyCode::N5, &[Modifier::LEFT_SHIFT], &[]);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.push_key(KeyCode::Y);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.push_key(KeyCode::Y);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.push_key(KeyCode::N5);
-            }
-
-            _ => {}
-        }
-    }
-
-    fn on_release(
-        &mut self,
-        prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-        match (prev.last_perm_layer, prev.layer) {
-            temp_layer!(KeymapLayer::Raise) => {
-                keyboard_report_state.pop_key(KeyCode::F4);
-            }
-            temp_layer!(KeymapLayer::LowerAnsi) => {
-                keyboard_report_state.pop_temp_modifiers();
-                keyboard_report_state.pop_key(KeyCode::N5);
-            }
-            temp_layer!(KeymapLayer::Lower | KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.pop_temp_modifiers();
-                keyboard_report_state.pop_key(KeyCode::N5);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.pop_key(KeyCode::Y);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.pop_key(KeyCode::Y);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.pop_key(KeyCode::N5);
-            }
-
-            _ => {}
-        }
-    }
-}
-
-impl KeyboardButton for LeftRow1Col0 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, _producer: &Producer) {
-        keyboard_report_state.push_key(KeyCode::ESCAPE);
-    }
-
-    fn on_release(
-        &mut self,
-        _prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-        keyboard_report_state.pop_key(KeyCode::ESCAPE);
-    }
-}
-
-impl KeyboardButton for LeftRow1Col1 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, _producer: &Producer) {
-        match (
-            keyboard_report_state.last_perm_layer,
-            keyboard_report_state.active_layer,
-        ) {
-            temp_layer!(KeymapLayer::Num) => {
-                keyboard_report_state.push_key(KeyCode::N1);
-            }
-            temp_layer!(KeymapLayer::Raise) => {
-                keyboard_report_state.push_key(KeyCode::LEFT_ARROW);
-            }
-            temp_layer!(KeymapLayer::LowerAnsi) => {
-                keyboard_report_state.push_key(KeyCode::KP_PLUS);
-            }
-            temp_layer!(KeymapLayer::Lower | KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.push_key(KeyCode::DASH);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.push_key(KeyCode::A);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.push_key(KeyCode::A);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.push_key(KeyCode::T);
-            }
-
-            _ => {}
-        }
-    }
-
-    fn on_release(
-        &mut self,
-        prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-        match (prev.last_perm_layer, prev.layer) {
-            temp_layer!(KeymapLayer::Num) => {
-                keyboard_report_state.pop_key(KeyCode::N1);
-            }
-            temp_layer!(KeymapLayer::Raise) => {
-                keyboard_report_state.pop_key(KeyCode::LEFT_ARROW);
-            }
-            temp_layer!(KeymapLayer::LowerAnsi) => {
-                keyboard_report_state.pop_key(KeyCode::KP_PLUS);
-            }
-            temp_layer!(KeymapLayer::Lower | KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.pop_key(KeyCode::DASH);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.pop_key(KeyCode::A);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.pop_key(KeyCode::A);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.pop_key(KeyCode::T);
-            }
-
-            _ => {}
-        }
-    }
-}
-impl KeyboardButton for LeftRow1Col2 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, _producer: &Producer) {
-        match (
-            keyboard_report_state.last_perm_layer,
-            keyboard_report_state.active_layer,
-        ) {
-            temp_layer!(KeymapLayer::Num) => {
-                keyboard_report_state.push_key(KeyCode::N2);
-            }
-            temp_layer!(KeymapLayer::Raise) => {
-                keyboard_report_state.push_key(KeyCode::RIGHT_ARROW);
-            }
-            temp_layer!(KeymapLayer::LowerAnsi) => {
-                keyboard_report_state.push_key(KeyCode::EQUALS);
-            }
-            temp_layer!(KeymapLayer::Lower | KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.temp_modify(KeyCode::N0, &[Modifier::LEFT_SHIFT], &[]);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.push_key(KeyCode::O);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.push_key(KeyCode::O);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.push_key(KeyCode::Q);
-            }
-
-            _ => {}
-        }
-    }
-
-    fn on_release(
-        &mut self,
-        prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-        match (prev.last_perm_layer, prev.layer) {
-            temp_layer!(KeymapLayer::Num) => {
-                keyboard_report_state.pop_key(KeyCode::N2);
-            }
-            temp_layer!(KeymapLayer::Raise) => {
-                keyboard_report_state.pop_key(KeyCode::RIGHT_ARROW);
-            }
-            temp_layer!(KeymapLayer::LowerAnsi) => {
-                keyboard_report_state.pop_key(KeyCode::EQUALS);
-            }
-            temp_layer!(KeymapLayer::Lower | KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.pop_temp_modifiers();
-                keyboard_report_state.pop_key(KeyCode::N0);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.pop_key(KeyCode::O);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.pop_key(KeyCode::O);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.pop_key(KeyCode::Q);
-            }
-
-            _ => {}
-        }
-    }
-}
-
-impl KeyboardButton for LeftRow1Col3 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, _producer: &Producer) {
-        match (
-            keyboard_report_state.last_perm_layer,
-            keyboard_report_state.active_layer,
-        ) {
-            temp_layer!(KeymapLayer::Num) => {
-                keyboard_report_state.push_key(KeyCode::N3);
-            }
-            temp_layer!(KeymapLayer::Raise) => {
-                keyboard_report_state.push_key(KeyCode::UP_ARROW);
-            }
-            temp_layer!(KeymapLayer::LowerAnsi) => {
-                keyboard_report_state.push_key(KeyCode::LEFT_BRACKET);
-            }
-            temp_layer!(KeymapLayer::Lower | KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.temp_modify(KeyCode::N8, &[Modifier::RIGHT_ALT], &[]);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.push_key(KeyCode::E);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.push_key(KeyCode::E);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.push_key(KeyCode::W);
-            }
-
-            _ => {}
-        }
-    }
-
-    fn on_release(
-        &mut self,
-        prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-        match (prev.last_perm_layer, prev.layer) {
-            temp_layer!(KeymapLayer::Num) => {
-                keyboard_report_state.pop_key(KeyCode::N3);
-            }
-            temp_layer!(KeymapLayer::Raise) => {
-                keyboard_report_state.pop_key(KeyCode::UP_ARROW);
-            }
-            temp_layer!(KeymapLayer::LowerAnsi) => {
-                keyboard_report_state.pop_key(KeyCode::LEFT_BRACKET);
-            }
-            temp_layer!(KeymapLayer::Lower | KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.pop_temp_modifiers();
-                keyboard_report_state.pop_key(KeyCode::N8);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.pop_key(KeyCode::E);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.pop_key(KeyCode::E);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.pop_key(KeyCode::W);
-            }
-
-            _ => {}
-        }
-    }
-}
-impl KeyboardButton for LeftRow1Col4 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, _producer: &Producer) {
-        match (
-            keyboard_report_state.last_perm_layer,
-            keyboard_report_state.active_layer,
-        ) {
-            temp_layer!(KeymapLayer::Num) => {
-                keyboard_report_state.push_key(KeyCode::N4);
-            }
-            temp_layer!(KeymapLayer::Raise) => {
-                keyboard_report_state.push_key(KeyCode::DOWN_ARROW);
-            }
-            temp_layer!(KeymapLayer::LowerAnsi) => {
-                keyboard_report_state.push_key(KeyCode::RIGHT_BRACKET);
-            }
-            temp_layer!(KeymapLayer::Lower | KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.temp_modify(KeyCode::N9, &[Modifier::RIGHT_ALT], &[]);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.push_key(KeyCode::U);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.push_key(KeyCode::U);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.push_key(KeyCode::E);
-            }
-
-            _ => {}
-        }
-    }
-
-    fn on_release(
-        &mut self,
-        prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-        match (prev.last_perm_layer, prev.layer) {
-            temp_layer!(KeymapLayer::Num) => {
-                keyboard_report_state.pop_key(KeyCode::N4);
-            }
-            temp_layer!(KeymapLayer::Raise) => {
-                keyboard_report_state.pop_key(KeyCode::DOWN_ARROW);
-            }
-            temp_layer!(KeymapLayer::LowerAnsi) => {
-                keyboard_report_state.pop_key(KeyCode::RIGHT_BRACKET);
-            }
-            temp_layer!(KeymapLayer::Lower | KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.pop_temp_modifiers();
-                keyboard_report_state.pop_key(KeyCode::N9);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.pop_key(KeyCode::U);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.pop_key(KeyCode::U);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.pop_key(KeyCode::E);
-            }
-
-            _ => {}
-        }
-    }
-}
-
-impl KeyboardButton for LeftRow1Col5 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, _producer: &Producer) {
-        match (
-            keyboard_report_state.last_perm_layer,
-            keyboard_report_state.active_layer,
-        ) {
-            temp_layer!(KeymapLayer::Num) => {
-                keyboard_report_state.push_key(KeyCode::N5);
-            }
-            temp_layer!(KeymapLayer::Raise) => {
-                keyboard_report_state.push_key(KeyCode::F11);
-            }
-            temp_layer!(KeymapLayer::LowerAnsi) => {
-                keyboard_report_state.temp_modify(KeyCode::SLASH, &[Modifier::LEFT_SHIFT], &[]);
-            }
-            temp_layer!(KeymapLayer::Lower | KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.temp_modify(KeyCode::DASH, &[Modifier::LEFT_SHIFT], &[]);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.push_key(KeyCode::I);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.push_key(KeyCode::I);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.push_key(KeyCode::R);
-            }
-
-            _ => {}
-        }
-    }
-
-    fn on_release(
-        &mut self,
-        prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-        match (prev.last_perm_layer, prev.layer) {
-            temp_layer!(KeymapLayer::Num) => {
-                keyboard_report_state.pop_key(KeyCode::N5);
-            }
-            temp_layer!(KeymapLayer::Raise) => {
-                keyboard_report_state.pop_key(KeyCode::F11);
-            }
-            temp_layer!(KeymapLayer::LowerAnsi) => {
-                keyboard_report_state.pop_temp_modifiers();
-                keyboard_report_state.pop_key(KeyCode::SLASH);
-            }
-            temp_layer!(KeymapLayer::Lower | KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.pop_temp_modifiers();
-                keyboard_report_state.pop_key(KeyCode::DASH);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.pop_key(KeyCode::I);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.pop_key(KeyCode::I);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.pop_key(KeyCode::R);
-            }
-
-            _ => {}
-        }
-    }
-}
-
-impl KeyboardButton for LeftRow2Col0 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, _producer: &Producer) {
-        keyboard_report_state.push_modifier(Modifier::LEFT_SHIFT);
-    }
-
-    fn on_release(
-        &mut self,
-        _prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-        keyboard_report_state.pop_modifier(Modifier::LEFT_SHIFT);
-    }
-}
-
-impl KeyboardButton for LeftRow2Col1 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, _producer: &Producer) {
-        match (
-            keyboard_report_state.last_perm_layer,
-            keyboard_report_state.active_layer,
-        ) {
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.push_key(KeyCode::SEMICOLON);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                if keyboard_report_state.has_modifier(Modifier::LEFT_SHIFT) {
-                    // Needs a shift, but that's already pressed
-                    keyboard_report_state.push_key(KeyCode::DOT);
-                    keyboard_report_state.jank.pressing_reg_colon = true;
-                } else {
-                    keyboard_report_state.temp_modify(KeyCode::COMMA, &[Modifier::LEFT_SHIFT], &[]);
-                    keyboard_report_state.jank.pressing_semicolon = true;
-                }
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.push_key(KeyCode::Y);
-            }
-            _ => {}
-        }
-    }
-
-    fn on_release(
-        &mut self,
-        prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-        match (prev.last_perm_layer, prev.layer) {
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.pop_key(KeyCode::SEMICOLON);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.pop_temp_modifiers();
-                if keyboard_report_state.jank.pressing_reg_colon {
-                    keyboard_report_state.pop_key(KeyCode::DOT);
-                    keyboard_report_state.jank.pressing_reg_colon = false;
-                }
-                if keyboard_report_state.jank.pressing_semicolon {
-                    keyboard_report_state.pop_key(KeyCode::COMMA);
-                    keyboard_report_state.jank.pressing_semicolon = false;
-                }
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.pop_key(KeyCode::Y);
-            }
-            _ => {}
-        }
-    }
-}
-
-impl KeyboardButton for LeftRow2Col2 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, _producer: &Producer) {
-        match (
-            keyboard_report_state.last_perm_layer,
-            keyboard_report_state.active_layer,
-        ) {
-            temp_layer!(KeymapLayer::LowerAnsi) => {
-                // Copy
-                keyboard_report_state.temp_modify(KeyCode::C, &[Modifier::LEFT_CONTROL], &[]);
-            }
-            temp_layer!(KeymapLayer::Lower | KeymapLayer::LowerSeMac) => {
-                // Copy
-                keyboard_report_state.temp_modify(KeyCode::C, &[Modifier::LEFT_CONTROL], &[]);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.push_key(KeyCode::Q);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.push_key(KeyCode::Q);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.push_key(KeyCode::A);
-            }
-
-            _ => {}
-        }
-    }
-
-    fn on_release(
-        &mut self,
-        prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-        match (prev.last_perm_layer, prev.layer) {
-            temp_layer!(KeymapLayer::LowerAnsi) => {
-                keyboard_report_state.pop_temp_modifiers();
-                keyboard_report_state.pop_key(KeyCode::C);
-            }
-            temp_layer!(KeymapLayer::Lower | KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.pop_temp_modifiers();
-                keyboard_report_state.pop_key(KeyCode::C);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.pop_key(KeyCode::Q);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.pop_key(KeyCode::Q);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.pop_key(KeyCode::A);
-            }
-
-            _ => {}
-        }
-    }
-}
-
-impl KeyboardButton for LeftRow2Col3 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, _producer: &Producer) {
-        match (
-            keyboard_report_state.last_perm_layer,
-            keyboard_report_state.active_layer,
-        ) {
-            temp_layer!(KeymapLayer::LowerAnsi) => {
-                keyboard_report_state.temp_modify(KeyCode::X, &[Modifier::LEFT_CONTROL], &[]);
-            }
-            temp_layer!(KeymapLayer::Lower | KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.temp_modify(KeyCode::X, &[Modifier::LEFT_CONTROL], &[]);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.push_key(KeyCode::J);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.push_key(KeyCode::J);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.push_key(KeyCode::S);
-            }
-
-            _ => {}
-        }
-    }
-
-    fn on_release(
-        &mut self,
-        prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-        match (prev.last_perm_layer, prev.layer) {
-            temp_layer!(KeymapLayer::LowerAnsi) => {
-                keyboard_report_state.pop_temp_modifiers();
-                keyboard_report_state.pop_key(KeyCode::X);
-            }
-            temp_layer!(KeymapLayer::Lower | KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.pop_temp_modifiers();
-                keyboard_report_state.pop_key(KeyCode::X);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.pop_key(KeyCode::J);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.pop_key(KeyCode::J);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.pop_key(KeyCode::S);
-            }
-
-            _ => {}
-        }
-    }
-}
-
-impl KeyboardButton for LeftRow2Col4 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, _producer: &Producer) {
-        match (
-            keyboard_report_state.last_perm_layer,
-            keyboard_report_state.active_layer,
-        ) {
-            temp_layer!(KeymapLayer::LowerAnsi) => {
-                keyboard_report_state.temp_modify(KeyCode::V, &[Modifier::LEFT_CONTROL], &[]);
-            }
-            temp_layer!(KeymapLayer::Lower | KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.temp_modify(KeyCode::V, &[Modifier::LEFT_CONTROL], &[]);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.push_key(KeyCode::K);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.push_key(KeyCode::K);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.push_key(KeyCode::D);
-            }
-
-            _ => {}
-        }
-    }
-
-    fn on_release(
-        &mut self,
-        prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-        match (prev.last_perm_layer, prev.layer) {
-            temp_layer!(KeymapLayer::LowerAnsi) => {
-                keyboard_report_state.pop_temp_modifiers();
-                keyboard_report_state.pop_key(KeyCode::V);
-            }
-            temp_layer!(KeymapLayer::Lower | KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.pop_temp_modifiers();
-                keyboard_report_state.pop_key(KeyCode::V);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.pop_key(KeyCode::K);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.pop_key(KeyCode::K);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.pop_key(KeyCode::D);
-            }
-
-            _ => {}
-        }
-    }
-}
-
-impl KeyboardButton for LeftRow2Col5 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, _producer: &Producer) {
-        match (
-            keyboard_report_state.last_perm_layer,
-            keyboard_report_state.active_layer,
-        ) {
-            temp_layer!(KeymapLayer::LowerAnsi) => {
-                keyboard_report_state.temp_modify(KeyCode::GRAVE, &[Modifier::LEFT_SHIFT], &[]);
-            }
-            temp_layer!(KeymapLayer::Lower) => {
-                // ~ Tilde double-tap to get it out immediately
-                keyboard_report_state.temp_modify(
-                    KeyCode::RIGHT_BRACKET,
-                    &[Modifier::RIGHT_ALT],
-                    &[],
-                );
-                keyboard_report_state.pop_key(KeyCode::RIGHT_BRACKET);
-                keyboard_report_state.temp_modify(
-                    KeyCode::RIGHT_BRACKET,
-                    &[Modifier::RIGHT_ALT],
-                    &[],
-                );
-            }
-            temp_layer!(KeymapLayer::LowerSeMac) => {
-                // ~ Tilde double-tap to get it out immediately
-                keyboard_report_state.temp_modify(
-                    KeyCode::RIGHT_BRACKET,
-                    &[Modifier::LEFT_ALT],
-                    &[],
-                );
-                keyboard_report_state.pop_key(KeyCode::RIGHT_BRACKET);
-                keyboard_report_state.temp_modify(KeyCode::SPACE, &[], &[Modifier::LEFT_ALT]);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.push_key(KeyCode::X);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.push_key(KeyCode::X);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.push_key(KeyCode::F);
-            }
-
-            _ => {}
-        }
-    }
-
-    fn on_release(
-        &mut self,
-        prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-        match (prev.last_perm_layer, prev.layer) {
-            temp_layer!(KeymapLayer::LowerAnsi) => {
-                keyboard_report_state.pop_temp_modifiers();
-                keyboard_report_state.pop_key(KeyCode::GRAVE);
-            }
-            temp_layer!(KeymapLayer::Lower) => {
-                keyboard_report_state.pop_temp_modifiers();
-                keyboard_report_state.pop_key(KeyCode::RIGHT_BRACKET);
-            }
-            temp_layer!(KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.pop_key(KeyCode::SPACE);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.pop_key(KeyCode::X);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.pop_key(KeyCode::X);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.pop_key(KeyCode::F);
-            }
-
-            _ => {}
-        }
-    }
-}
-
-impl KeyboardButton for LeftRow3Col0 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, _producer: &Producer) {
-        keyboard_report_state.push_modifier(Modifier::LEFT_CONTROL);
-    }
-
-    fn on_release(
-        &mut self,
-        _prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-        keyboard_report_state.pop_modifier(Modifier::LEFT_CONTROL);
-    }
-}
-
-impl KeyboardButton for LeftRow3Col1 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, _producer: &Producer) {
-        match (
-            keyboard_report_state.last_perm_layer,
-            keyboard_report_state.active_layer,
-        ) {
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.push_modifier(Modifier::LEFT_GUI);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.push_modifier(Modifier::LEFT_GUI);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.push_key(KeyCode::Z);
-            }
-            _ => {}
-        }
-    }
-
-    fn on_release(
-        &mut self,
-        prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-        match (prev.last_perm_layer, prev.layer) {
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.pop_modifier(Modifier::LEFT_GUI);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.pop_modifier(Modifier::LEFT_GUI);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.pop_key(KeyCode::Z);
-            }
-            _ => {}
-        }
-    }
-}
-
-impl KeyboardButton for LeftRow3Col2 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, _producer: &Producer) {
-        match (
-            keyboard_report_state.last_perm_layer,
-            keyboard_report_state.active_layer,
-        ) {
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.push_modifier(Modifier::LEFT_ALT);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.push_modifier(Modifier::LEFT_ALT);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.push_key(KeyCode::X);
-            }
-            _ => {}
-        }
-    }
-
-    fn on_release(
-        &mut self,
-        prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-        match (prev.last_perm_layer, prev.layer) {
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.pop_modifier(Modifier::LEFT_ALT);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.pop_modifier(Modifier::LEFT_ALT);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.pop_key(KeyCode::X);
-            }
-            _ => {}
-        }
-    }
-}
-
-impl KeyboardButton for LeftRow3Col3 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, _producer: &Producer) {
-        match (
-            keyboard_report_state.last_perm_layer,
-            keyboard_report_state.active_layer,
-        ) {
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.push_key(KeyCode::C);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {}
-            base_layer!(KeymapLayer::DvorakAnsi) => {}
-            _ => {}
-        }
-    }
-
-    fn on_release(
-        &mut self,
-        prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-        match (prev.last_perm_layer, prev.layer) {
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.pop_key(KeyCode::C);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {}
-            base_layer!(KeymapLayer::DvorakAnsi) => {}
-            _ => {}
-        }
-    }
-}
-
-impl KeyboardButton for LeftRow3Col4 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, producer: &Producer) {
-        match (
-            keyboard_report_state.last_perm_layer,
-            keyboard_report_state.active_layer,
-        ) {
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.push_layer_with_fallback(KeymapLayer::LowerAnsi);
-                push_layer_change(producer, KeymapLayer::DvorakAnsi);
-            }
-            base_layer!(KeymapLayer::DvorakSe) => {
-                keyboard_report_state.push_layer_with_fallback(KeymapLayer::Lower);
-                push_layer_change(producer, KeymapLayer::DvorakSe);
-            }
-            base_layer!(KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.push_layer_with_fallback(KeymapLayer::LowerSeMac);
-                push_layer_change(producer, KeymapLayer::DvorakSeMac);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.push_key(KeyCode::V);
-            }
-            _ => {}
-        }
-    }
-
-    fn on_release(
-        &mut self,
-        prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        producer: &Producer,
-    ) {
-        match (prev.last_perm_layer, prev.layer) {
-            temp_layer!(KeymapLayer::LowerAnsi) => {
-                keyboard_report_state.pop_layer(KeymapLayer::LowerAnsi);
-                push_layer_change(producer, keyboard_report_state.active_layer);
-            }
-            temp_layer!(KeymapLayer::Lower) => {
-                keyboard_report_state.pop_layer(KeymapLayer::Lower);
-                push_layer_change(producer, keyboard_report_state.active_layer);
-            }
-            temp_layer!(KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.pop_layer(KeymapLayer::LowerSeMac);
-                push_layer_change(producer, keyboard_report_state.active_layer);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.pop_key(KeyCode::V);
-            }
-
-            base_layer!(KeymapLayer::DvorakAnsi) => {}
-            base_layer!(KeymapLayer::DvorakSe) => {}
-            _ => {}
-        }
-    }
-}
-
-impl KeyboardButton for LeftRow3Col5 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, _producer: &Producer) {
-        match (
-            keyboard_report_state.last_perm_layer,
-            keyboard_report_state.active_layer,
-        ) {
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.push_key(KeyCode::SPACE);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.push_key(KeyCode::SPACE);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.push_key(KeyCode::B);
-            }
-            _ => {}
-        }
-    }
-
-    fn on_release(
-        &mut self,
-        prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-        match (prev.last_perm_layer, prev.layer) {
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.pop_key(KeyCode::SPACE);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.pop_key(KeyCode::SPACE);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.pop_key(KeyCode::B);
-            }
-            _ => {}
-        }
-    }
-}
-
-// Row 4 col 0 does not exist
-impl KeyboardButton for LeftRow4Col1 {
-    fn on_press(&mut self, _keyboard_report_state: &mut KeyboardReportState, producer: &Producer) {
-        push_reboot_and_halt(producer);
-    }
-
-    fn on_release(
-        &mut self,
-        _prev: LastPressState,
-        _keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-    }
-}
-
-impl KeyboardButton for LeftRow4Col2 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, _producer: &Producer) {
-        match (
-            keyboard_report_state.last_perm_layer,
-            keyboard_report_state.active_layer,
-        ) {
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.push_modifier(Modifier::LEFT_GUI);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {}
-            base_layer!(KeymapLayer::DvorakAnsi) => {}
-            _ => {}
-        }
-    }
-
-    fn on_release(
-        &mut self,
-        prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-        match (prev.last_perm_layer, prev.layer) {
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.pop_modifier(Modifier::LEFT_GUI);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {}
-            base_layer!(KeymapLayer::DvorakAnsi) => {}
-            _ => {}
-        }
-    }
-}
-
-impl KeyboardButton for LeftRow4Col3 {
-    fn on_press(&mut self, _keyboard_report_state: &mut KeyboardReportState, _producer: &Producer) {
-    }
-
-    fn on_release(
-        &mut self,
-        _prev: LastPressState,
-        _keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-    }
-}
-
-impl KeyboardButton for LeftRow4Col4 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, _producer: &Producer) {
-        match (
-            keyboard_report_state.last_perm_layer,
-            keyboard_report_state.active_layer,
-        ) {
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.push_key(KeyCode::SPACE);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                // _
-                keyboard_report_state.temp_modify(KeyCode::SLASH, &[Modifier::LEFT_SHIFT], &[]);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {}
-            _ => {}
-        }
-    }
-
-    fn on_release(
-        &mut self,
-        prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-        match (prev.last_perm_layer, prev.layer) {
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.pop_key(KeyCode::SPACE);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.pop_key(KeyCode::SLASH);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {}
-            _ => {}
-        }
-    }
-}
-
-impl KeyboardButton for LeftRow4Col5 {
-    fn on_press(&mut self, _keyboard_report_state: &mut KeyboardReportState, _producer: &Producer) {
-    }
-    fn on_release(
-        &mut self,
-        _prev: LastPressState,
-        _keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-    }
-}
-
-// Right side, goes from right to left
-
-impl KeyboardButton for RightRow0Col0 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, _producer: &Producer) {
-        keyboard_report_state.push_key(KeyCode::BACKSPACE);
-    }
-
-    fn on_release(
-        &mut self,
-        _prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-        keyboard_report_state.pop_key(KeyCode::BACKSPACE);
-    }
-}
-
-impl KeyboardButton for RightRow0Col1 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, _producer: &Producer) {
-        match (
-            keyboard_report_state.last_perm_layer,
-            keyboard_report_state.active_layer,
-        ) {
-            temp_layer!(KeymapLayer::Raise) => {
-                keyboard_report_state.push_key(KeyCode::F10);
-            }
-            temp_layer!(KeymapLayer::LowerAnsi) => {
-                keyboard_report_state.temp_modify(KeyCode::BACKSLASH, &[Modifier::LEFT_SHIFT], &[]);
-            }
-            temp_layer!(KeymapLayer::Lower | KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.temp_modify(KeyCode::BACKSLASH, &[Modifier::LEFT_SHIFT], &[]);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.push_key(KeyCode::L);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.push_key(KeyCode::L);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.push_key(KeyCode::N0);
-            }
-
-            _ => {}
-        }
-    }
-
-    fn on_release(
-        &mut self,
-        prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-        match (prev.last_perm_layer, prev.layer) {
-            temp_layer!(KeymapLayer::Raise) => {
-                keyboard_report_state.pop_key(KeyCode::F10);
-            }
-            temp_layer!(KeymapLayer::LowerAnsi) => {
-                keyboard_report_state.pop_temp_modifiers();
-                keyboard_report_state.pop_key(KeyCode::N8);
-            }
-            temp_layer!(KeymapLayer::Lower | KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.pop_temp_modifiers();
-                keyboard_report_state.pop_key(KeyCode::BACKSLASH);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.pop_key(KeyCode::L);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.pop_key(KeyCode::L);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.pop_key(KeyCode::N0);
-            }
-
-            _ => {}
-        }
-    }
-}
-
-impl KeyboardButton for RightRow0Col2 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, _producer: &Producer) {
-        match (
-            keyboard_report_state.last_perm_layer,
-            keyboard_report_state.active_layer,
-        ) {
-            temp_layer!(KeymapLayer::Raise) => {
-                keyboard_report_state.push_key(KeyCode::F9);
-            }
-            temp_layer!(KeymapLayer::LowerAnsi) => {
-                keyboard_report_state.temp_modify(KeyCode::N8, &[Modifier::LEFT_SHIFT], &[]);
-            }
-            temp_layer!(KeymapLayer::Lower | KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.temp_modify(KeyCode::N9, &[Modifier::LEFT_SHIFT], &[]);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.push_key(KeyCode::R);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.push_key(KeyCode::R);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.push_key(KeyCode::N9);
-            }
-
-            _ => {}
-        }
-    }
-
-    fn on_release(
-        &mut self,
-        prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-        match (prev.last_perm_layer, prev.layer) {
-            temp_layer!(KeymapLayer::Raise) => {
-                keyboard_report_state.pop_key(KeyCode::F9);
-            }
-            temp_layer!(KeymapLayer::LowerAnsi) => {
-                keyboard_report_state.pop_temp_modifiers();
-                keyboard_report_state.pop_key(KeyCode::N0);
-            }
-            temp_layer!(KeymapLayer::Lower | KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.pop_temp_modifiers();
-                keyboard_report_state.pop_key(KeyCode::N9);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.pop_key(KeyCode::R);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.pop_key(KeyCode::R);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.pop_key(KeyCode::N9);
-            }
-
-            _ => {}
-        }
-    }
-}
-
-impl KeyboardButton for RightRow0Col3 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, _producer: &Producer) {
-        match (
-            keyboard_report_state.last_perm_layer,
-            keyboard_report_state.active_layer,
-        ) {
-            temp_layer!(KeymapLayer::Raise) => {
-                keyboard_report_state.push_key(KeyCode::F8);
-            }
-            temp_layer!(KeymapLayer::LowerAnsi) => {
-                keyboard_report_state.temp_modify(KeyCode::N9, &[Modifier::LEFT_SHIFT], &[]);
-            }
-            temp_layer!(KeymapLayer::Lower | KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.temp_modify(KeyCode::N8, &[Modifier::LEFT_SHIFT], &[]);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.push_key(KeyCode::C);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.push_key(KeyCode::C);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.push_key(KeyCode::N8);
-            }
-
-            _ => {}
-        }
-    }
-
-    fn on_release(
-        &mut self,
-        prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-        match (prev.last_perm_layer, prev.layer) {
-            temp_layer!(KeymapLayer::Raise) => {
-                keyboard_report_state.pop_key(KeyCode::F8);
-            }
-            temp_layer!(KeymapLayer::LowerAnsi) => {
-                keyboard_report_state.pop_temp_modifiers();
-                keyboard_report_state.pop_key(KeyCode::N9);
-            }
-            temp_layer!(KeymapLayer::Lower | KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.pop_temp_modifiers();
-                keyboard_report_state.pop_key(KeyCode::N8);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.pop_key(KeyCode::C);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.pop_key(KeyCode::C);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.pop_key(KeyCode::N8);
-            }
-
-            _ => {}
-        }
-    }
-}
-
-impl KeyboardButton for RightRow0Col4 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, _producer: &Producer) {
-        match (
-            keyboard_report_state.last_perm_layer,
-            keyboard_report_state.active_layer,
-        ) {
-            temp_layer!(KeymapLayer::Raise) => {
-                keyboard_report_state.push_key(KeyCode::F7);
-            }
-            temp_layer!(KeymapLayer::LowerAnsi) => {
-                keyboard_report_state.temp_modify(KeyCode::N7, &[Modifier::LEFT_SHIFT], &[]);
-            }
-            temp_layer!(KeymapLayer::Lower | KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.temp_modify(KeyCode::N6, &[Modifier::LEFT_SHIFT], &[]);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.push_key(KeyCode::G);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.push_key(KeyCode::G);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.push_key(KeyCode::N7);
-            }
-
-            _ => {}
-        }
-    }
-
-    fn on_release(
-        &mut self,
-        prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-        match (prev.last_perm_layer, prev.layer) {
-            temp_layer!(KeymapLayer::Raise) => {
-                keyboard_report_state.pop_key(KeyCode::F7);
-            }
-            temp_layer!(KeymapLayer::LowerAnsi) => {
-                keyboard_report_state.pop_temp_modifiers();
-                keyboard_report_state.pop_key(KeyCode::N7);
-            }
-            temp_layer!(KeymapLayer::Lower | KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.pop_temp_modifiers();
-                keyboard_report_state.pop_key(KeyCode::N6);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.pop_key(KeyCode::G);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.pop_key(KeyCode::G);
-            }
-
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.pop_key(KeyCode::N7);
-            }
-
-            _ => {}
-        }
-    }
-}
-
-impl KeyboardButton for RightRow0Col5 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, _producer: &Producer) {
-        match (
-            keyboard_report_state.last_perm_layer,
-            keyboard_report_state.active_layer,
-        ) {
-            temp_layer!(KeymapLayer::Raise) => {
-                keyboard_report_state.push_key(KeyCode::F6);
-            }
-            temp_layer!(KeymapLayer::LowerAnsi) => {
-                keyboard_report_state.temp_modify(KeyCode::N6, &[Modifier::LEFT_SHIFT], &[]);
-            }
-            temp_layer!(KeymapLayer::Lower) => {
-                // Double-tap to get ^ on one press, not like I ever use circ for anything else
-                if keyboard_report_state.has_modifier(Modifier::ANY_SHIFT) {
-                    keyboard_report_state.push_key(KeyCode::RIGHT_BRACKET);
-                    keyboard_report_state.pop_key(KeyCode::RIGHT_BRACKET);
-                    keyboard_report_state.push_key(KeyCode::RIGHT_BRACKET);
-                } else {
-                    keyboard_report_state.temp_modify(
-                        KeyCode::RIGHT_BRACKET,
-                        &[Modifier::LEFT_SHIFT],
-                        &[],
-                    );
-                    keyboard_report_state.pop_key(KeyCode::RIGHT_BRACKET);
-                    keyboard_report_state.temp_modify(
-                        KeyCode::RIGHT_BRACKET,
-                        &[Modifier::LEFT_SHIFT],
-                        &[],
-                    );
-                }
-            }
-            temp_layer!(KeymapLayer::LowerSeMac) => {
-                // macos circ (^)
-                keyboard_report_state.temp_modify(
-                    KeyCode::RIGHT_BRACKET,
-                    &[Modifier::LEFT_SHIFT],
-                    &[],
-                );
-                keyboard_report_state.pop_key(KeyCode::RIGHT_BRACKET);
-                keyboard_report_state.temp_modify(KeyCode::SPACE, &[], &[Modifier::LEFT_SHIFT]);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.push_key(KeyCode::F);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.push_key(KeyCode::F);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.push_key(KeyCode::N6);
-            }
-
-            _ => {}
-        }
-    }
-
-    fn on_release(
-        &mut self,
-        prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-        match (prev.last_perm_layer, prev.layer) {
-            temp_layer!(KeymapLayer::Raise) => {
-                keyboard_report_state.pop_key(KeyCode::F6);
-            }
-            temp_layer!(KeymapLayer::LowerAnsi) => {
-                keyboard_report_state.pop_temp_modifiers();
-                keyboard_report_state.pop_key(KeyCode::N6);
-            }
-            temp_layer!(KeymapLayer::Lower) => {
-                keyboard_report_state.pop_temp_modifiers();
-                keyboard_report_state.pop_key(KeyCode::RIGHT_BRACKET);
-            }
-            temp_layer!(KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.pop_temp_modifiers();
-                keyboard_report_state.pop_key(KeyCode::SPACE);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.pop_key(KeyCode::F);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.pop_key(KeyCode::F);
-            }
-
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.pop_key(KeyCode::N6);
-            }
-
-            _ => {}
-        }
-    }
-}
-
-impl KeyboardButton for RightRow1Col0 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, _producer: &Producer) {
-        keyboard_report_state.push_key(KeyCode::ENTER);
-    }
-
-    fn on_release(
-        &mut self,
-        _prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-        keyboard_report_state.pop_key(KeyCode::ENTER);
-    }
-}
-
-impl KeyboardButton for RightRow1Col1 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, _producer: &Producer) {
-        match (
-            keyboard_report_state.last_perm_layer,
-            keyboard_report_state.active_layer,
-        ) {
-            temp_layer!(KeymapLayer::Num) => {
-                keyboard_report_state.push_key(KeyCode::N0);
-            }
-            temp_layer!(KeymapLayer::Raise) => {
-                keyboard_report_state.push_key(KeyCode::KC_DELF);
-            }
-            temp_layer!(KeymapLayer::LowerAnsi) => {
-                keyboard_report_state.push_key(KeyCode::DASH);
-            }
-            temp_layer!(KeymapLayer::Lower | KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.push_key(KeyCode::SLASH);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.push_key(KeyCode::S);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.push_key(KeyCode::S);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.push_key(KeyCode::SEMICOLON);
-            }
-
-            _ => {}
-        }
-    }
-
-    fn on_release(
-        &mut self,
-        prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-        match (prev.last_perm_layer, prev.layer) {
-            temp_layer!(KeymapLayer::Num) => {
-                keyboard_report_state.pop_key(KeyCode::N0);
-            }
-            temp_layer!(KeymapLayer::Raise) => {
-                keyboard_report_state.pop_key(KeyCode::KC_DELF);
-            }
-            temp_layer!(KeymapLayer::LowerAnsi) => {
-                keyboard_report_state.pop_key(KeyCode::DASH);
-            }
-            temp_layer!(KeymapLayer::Lower | KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.pop_key(KeyCode::SLASH);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.pop_key(KeyCode::S);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.pop_key(KeyCode::S);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.pop_key(KeyCode::SEMICOLON);
-            }
-
-            _ => {}
-        }
-    }
-}
-
-impl KeyboardButton for RightRow1Col2 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, _producer: &Producer) {
-        match (
-            keyboard_report_state.last_perm_layer,
-            keyboard_report_state.active_layer,
-        ) {
-            temp_layer!(KeymapLayer::Num) => {
-                keyboard_report_state.push_key(KeyCode::N9);
-            }
-            temp_layer!(KeymapLayer::Raise) => {
-                keyboard_report_state.push_key(KeyCode::HOME);
-            }
-            temp_layer!(KeymapLayer::LowerAnsi) => {
-                keyboard_report_state.temp_modify(
-                    KeyCode::RIGHT_BRACKET,
-                    &[Modifier::RIGHT_ALT],
-                    &[],
-                );
-            }
-            temp_layer!(KeymapLayer::Lower) => {
-                keyboard_report_state.temp_modify(KeyCode::N0, &[Modifier::RIGHT_ALT], &[]);
-            }
-            temp_layer!(KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.temp_modify(
-                    KeyCode::N9,
-                    &[Modifier::LEFT_SHIFT, Modifier::LEFT_ALT],
-                    &[],
-                );
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.push_key(KeyCode::N);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.push_key(KeyCode::N);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.push_key(KeyCode::L);
-            }
-
-            _ => {}
-        }
-    }
-
-    fn on_release(
-        &mut self,
-        prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-        match (prev.last_perm_layer, prev.layer) {
-            temp_layer!(KeymapLayer::Num) => {
-                keyboard_report_state.pop_key(KeyCode::N9);
-            }
-            temp_layer!(KeymapLayer::Raise) => {
-                keyboard_report_state.pop_key(KeyCode::HOME);
-            }
-            temp_layer!(KeymapLayer::LowerAnsi) => {
-                keyboard_report_state.pop_temp_modifiers();
-                keyboard_report_state.pop_key(KeyCode::RIGHT_BRACKET);
-            }
-            temp_layer!(KeymapLayer::Lower) => {
-                keyboard_report_state.pop_temp_modifiers();
-                keyboard_report_state.pop_key(KeyCode::N0);
-            }
-            temp_layer!(KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.pop_temp_modifiers();
-                keyboard_report_state.pop_key(KeyCode::N9);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.pop_key(KeyCode::N);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.pop_key(KeyCode::N);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.pop_key(KeyCode::L);
-            }
-
-            _ => {}
-        }
-    }
-}
-
-impl KeyboardButton for RightRow1Col3 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, _producer: &Producer) {
-        match (
-            keyboard_report_state.last_perm_layer,
-            keyboard_report_state.active_layer,
-        ) {
-            temp_layer!(KeymapLayer::Num) => {
-                keyboard_report_state.push_key(KeyCode::N8);
-            }
-            temp_layer!(KeymapLayer::Raise) => {
-                keyboard_report_state.push_key(KeyCode::PAGE_UP);
-            }
-            temp_layer!(KeymapLayer::LowerAnsi) => {
-                keyboard_report_state.temp_modify(
-                    KeyCode::LEFT_BRACKET,
-                    &[Modifier::LEFT_SHIFT],
-                    &[],
-                );
-            }
-            temp_layer!(KeymapLayer::Lower) => {
-                keyboard_report_state.temp_modify(KeyCode::N7, &[Modifier::RIGHT_ALT], &[]);
-            }
-            temp_layer!(KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.temp_modify(
-                    KeyCode::N8,
-                    &[Modifier::LEFT_SHIFT, Modifier::LEFT_ALT],
-                    &[],
-                );
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.push_key(KeyCode::T);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.push_key(KeyCode::T);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.push_key(KeyCode::K);
-            }
-
-            _ => {}
-        }
-    }
-
-    fn on_release(
-        &mut self,
-        prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-        match (prev.last_perm_layer, prev.layer) {
-            temp_layer!(KeymapLayer::Num) => {
-                keyboard_report_state.pop_key(KeyCode::N8);
-            }
-            temp_layer!(KeymapLayer::Raise) => {
-                keyboard_report_state.pop_key(KeyCode::PAGE_UP);
-            }
-            temp_layer!(KeymapLayer::LowerAnsi) => {
-                keyboard_report_state.pop_temp_modifiers();
-                keyboard_report_state.pop_key(KeyCode::LEFT_BRACKET);
-            }
-            temp_layer!(KeymapLayer::Lower) => {
-                keyboard_report_state.pop_temp_modifiers();
-                keyboard_report_state.pop_key(KeyCode::N7);
-            }
-            temp_layer!(KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.pop_temp_modifiers();
-                keyboard_report_state.pop_key(KeyCode::N8);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.pop_key(KeyCode::T);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.pop_key(KeyCode::T);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.pop_key(KeyCode::K);
-            }
-
-            _ => {}
-        }
-    }
-}
-
-impl KeyboardButton for RightRow1Col4 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, _producer: &Producer) {
-        match (
-            keyboard_report_state.last_perm_layer,
-            keyboard_report_state.active_layer,
-        ) {
-            temp_layer!(KeymapLayer::Num) => {
-                keyboard_report_state.push_key(KeyCode::N7);
-            }
-            temp_layer!(KeymapLayer::Raise) => {
-                keyboard_report_state.push_key(KeyCode::PRINT_SCREEN);
-            }
-            temp_layer!(KeymapLayer::LowerAnsi) => {
-                keyboard_report_state.push_key(KeyCode::SLASH);
-            }
-            temp_layer!(KeymapLayer::Lower | KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.temp_modify(KeyCode::N7, &[Modifier::LEFT_SHIFT], &[]);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.push_key(KeyCode::H);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.push_key(KeyCode::H);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.push_key(KeyCode::J);
-            }
-
-            _ => {}
-        }
-    }
-
-    fn on_release(
-        &mut self,
-        prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-        match (prev.last_perm_layer, prev.layer) {
-            temp_layer!(KeymapLayer::Lower | KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.pop_temp_modifiers();
-                keyboard_report_state.pop_key(KeyCode::N7);
-            }
-            temp_layer!(KeymapLayer::Num) => {
-                keyboard_report_state.pop_key(KeyCode::N7);
-            }
-            temp_layer!(KeymapLayer::Raise) => {
-                keyboard_report_state.pop_key(KeyCode::PRINT_SCREEN);
-            }
-            temp_layer!(KeymapLayer::LowerAnsi) => {
-                keyboard_report_state.pop_key(KeyCode::SLASH);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.pop_key(KeyCode::H);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.pop_key(KeyCode::H);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.pop_key(KeyCode::J);
-            }
-
-            _ => {}
-        }
-    }
-}
-
-impl KeyboardButton for RightRow1Col5 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, _producer: &Producer) {
-        match (
-            keyboard_report_state.last_perm_layer,
-            keyboard_report_state.active_layer,
-        ) {
-            temp_layer!(KeymapLayer::Num) => {
-                keyboard_report_state.push_key(KeyCode::N6);
-            }
-            temp_layer!(KeymapLayer::Raise) => {
-                keyboard_report_state.push_key(KeyCode::F12);
-            }
-            temp_layer!(KeymapLayer::LowerAnsi) => {
-                keyboard_report_state.push_key(KeyCode::BACKSLASH);
-            }
-            temp_layer!(KeymapLayer::Lower) => {
-                keyboard_report_state.temp_modify(KeyCode::DASH, &[Modifier::RIGHT_ALT], &[]);
-            }
-            temp_layer!(KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.temp_modify(
-                    KeyCode::N7,
-                    &[Modifier::LEFT_SHIFT, Modifier::LEFT_ALT],
-                    &[],
-                );
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.push_key(KeyCode::D);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.push_key(KeyCode::D);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.push_key(KeyCode::H);
-            }
-
-            _ => {}
-        }
-    }
-
-    fn on_release(
-        &mut self,
-        prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-        match (prev.last_perm_layer, prev.layer) {
-            temp_layer!(KeymapLayer::Num) => {
-                keyboard_report_state.pop_key(KeyCode::N6);
-            }
-            temp_layer!(KeymapLayer::Raise) => {
-                keyboard_report_state.pop_key(KeyCode::F12);
-            }
-            temp_layer!(KeymapLayer::LowerAnsi) => {
-                keyboard_report_state.pop_key(KeyCode::BACKSLASH);
-            }
-            temp_layer!(KeymapLayer::Lower) => {
-                keyboard_report_state.pop_temp_modifiers();
-                keyboard_report_state.pop_key(KeyCode::DASH);
-            }
-            temp_layer!(KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.pop_temp_modifiers();
-                keyboard_report_state.pop_key(KeyCode::N7);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.pop_key(KeyCode::D);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.pop_key(KeyCode::D);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.pop_key(KeyCode::H);
-            }
-
-            _ => {}
-        }
-    }
-}
-
-impl KeyboardButton for RightRow2Col0 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, _producer: &Producer) {
-        keyboard_report_state.push_modifier(Modifier::LEFT_SHIFT);
-    }
-
-    fn on_release(
-        &mut self,
-        _prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-        keyboard_report_state.pop_modifier(Modifier::LEFT_SHIFT);
-    }
-}
-
-impl KeyboardButton for RightRow2Col1 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, _producer: &Producer) {
-        match (
-            keyboard_report_state.last_perm_layer,
-            keyboard_report_state.active_layer,
-        ) {
-            temp_layer!(KeymapLayer::Raise) => {
-                keyboard_report_state.push_key(KeyCode::INSERT);
-            }
-            temp_layer!(KeymapLayer::Lower | KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.push_key(KeyCode::SEMICOLON);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.push_key(KeyCode::Z);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.push_key(KeyCode::Z);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.push_key(KeyCode::QUOTE);
-            }
-
-            _ => {}
-        }
-    }
-
-    fn on_release(
-        &mut self,
-        prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-        match (prev.last_perm_layer, prev.layer) {
-            temp_layer!(KeymapLayer::Raise) => {
-                keyboard_report_state.pop_key(KeyCode::INSERT);
-            }
-            temp_layer!(KeymapLayer::Lower | KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.pop_key(KeyCode::SEMICOLON);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.pop_key(KeyCode::Z);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.pop_key(KeyCode::Z);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.pop_key(KeyCode::QUOTE);
-            }
-
-            _ => {}
-        }
-    }
-}
-
-impl KeyboardButton for RightRow2Col2 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, producer: &Producer) {
-        match (
-            keyboard_report_state.last_perm_layer,
-            keyboard_report_state.active_layer,
-        ) {
-            temp_layer!(KeymapLayer::Settings) => {
-                keyboard_report_state.set_perm_layer(KeymapLayer::DvorakAnsi);
-                push_layer_change(producer, keyboard_report_state.active_layer);
-            }
-            temp_layer!(KeymapLayer::Raise) => {
-                keyboard_report_state.push_key(KeyCode::END);
-            }
-            temp_layer!(KeymapLayer::Lower | KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.push_key(KeyCode::QUOTE);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.push_key(KeyCode::V);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.push_key(KeyCode::V);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.push_key(KeyCode::DOT);
-            }
-
-            _ => {}
-        }
-    }
-
-    fn on_release(
-        &mut self,
-        prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-        match (prev.last_perm_layer, prev.layer) {
-            temp_layer!(KeymapLayer::Raise) => {
-                keyboard_report_state.pop_key(KeyCode::END);
-            }
-            temp_layer!(KeymapLayer::Lower | KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.pop_key(KeyCode::QUOTE);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.pop_key(KeyCode::V);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.pop_key(KeyCode::V);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.pop_key(KeyCode::DOT);
-            }
-
-            _ => {}
-        }
-    }
-}
-
-impl KeyboardButton for RightRow2Col3 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, producer: &Producer) {
-        match (
-            keyboard_report_state.last_perm_layer,
-            keyboard_report_state.active_layer,
-        ) {
-            temp_layer!(KeymapLayer::Settings) => {
-                keyboard_report_state.set_perm_layer(KeymapLayer::DvorakSe);
-                push_layer_change(producer, keyboard_report_state.active_layer);
-            }
-            temp_layer!(KeymapLayer::Raise) => {
-                keyboard_report_state.push_key(KeyCode::PAGE_DOWN);
-            }
-            temp_layer!(KeymapLayer::Lower | KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.push_key(KeyCode::LEFT_BRACKET);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.push_key(KeyCode::W);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.push_key(KeyCode::W);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.push_key(KeyCode::COMMA);
-            }
-
-            _ => {}
-        }
-    }
-
-    fn on_release(
-        &mut self,
-        prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-        match (prev.last_perm_layer, prev.layer) {
-            temp_layer!(KeymapLayer::Raise) => {
-                keyboard_report_state.pop_key(KeyCode::PAGE_DOWN);
-            }
-            temp_layer!(KeymapLayer::Lower | KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.pop_key(KeyCode::LEFT_BRACKET);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.pop_key(KeyCode::W);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.pop_key(KeyCode::W);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.pop_key(KeyCode::COMMA);
-            }
-
-            _ => {}
-        }
-    }
-}
-
-impl KeyboardButton for RightRow2Col4 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, producer: &Producer) {
-        match (
-            keyboard_report_state.last_perm_layer,
-            keyboard_report_state.active_layer,
-        ) {
-            temp_layer!(KeymapLayer::Settings) => {
-                keyboard_report_state.set_perm_layer(KeymapLayer::QwertyGaming);
-                push_layer_change(producer, keyboard_report_state.active_layer);
-            }
-            temp_layer!(KeymapLayer::LowerAnsi) => {
-                keyboard_report_state.push_key(KeyCode::PIPE);
-            }
-            temp_layer!(KeymapLayer::Lower) => {
-                keyboard_report_state.temp_modify(
-                    KeyCode::NON_US_BACKSLASH,
-                    &[Modifier::RIGHT_ALT],
-                    &[],
-                );
-            }
-            temp_layer!(KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.temp_modify(KeyCode::N7, &[Modifier::LEFT_ALT], &[]);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.push_key(KeyCode::M);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.push_key(KeyCode::M);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.push_key(KeyCode::M);
-            }
-
-            _ => {}
-        }
-    }
-
-    fn on_release(
-        &mut self,
-        prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-        match (prev.last_perm_layer, prev.layer) {
-            temp_layer!(KeymapLayer::LowerAnsi) => {
-                keyboard_report_state.pop_key(KeyCode::PIPE);
-            }
-            temp_layer!(KeymapLayer::Lower) => {
-                keyboard_report_state.pop_temp_modifiers();
-                keyboard_report_state.pop_key(KeyCode::NON_US_BACKSLASH);
-            }
-            temp_layer!(KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.pop_temp_modifiers();
-                keyboard_report_state.pop_key(KeyCode::N7);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.pop_key(KeyCode::M);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.pop_key(KeyCode::M);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.pop_key(KeyCode::M);
-            }
-
-            _ => {}
-        }
-    }
-}
-
-impl KeyboardButton for RightRow2Col5 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, producer: &Producer) {
-        match (
-            keyboard_report_state.last_perm_layer,
-            keyboard_report_state.active_layer,
-        ) {
-            temp_layer!(KeymapLayer::Settings) => {
-                keyboard_report_state.set_perm_layer(KeymapLayer::DvorakSeMac);
-                push_layer_change(producer, keyboard_report_state.active_layer);
-            }
-            temp_layer!(KeymapLayer::LowerAnsi) => {
-                keyboard_report_state.push_key(KeyCode::GRAVE);
-            }
-            temp_layer!(KeymapLayer::Lower) => {
-                if keyboard_report_state.has_modifier(Modifier::ANY_SHIFT) {
-                    keyboard_report_state.push_key(KeyCode::EQUALS);
-                } else {
-                    keyboard_report_state.temp_modify(
-                        KeyCode::EQUALS,
-                        &[Modifier::LEFT_SHIFT],
-                        &[],
-                    );
-                    keyboard_report_state.pop_key(KeyCode::EQUALS);
-                    keyboard_report_state.temp_modify(
-                        KeyCode::EQUALS,
-                        &[Modifier::LEFT_SHIFT],
-                        &[],
-                    );
-                }
-            }
-            temp_layer!(KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.temp_modify(KeyCode::EQUALS, &[Modifier::LEFT_SHIFT], &[]);
-                keyboard_report_state.pop_key(KeyCode::EQUALS);
-                keyboard_report_state.temp_modify(KeyCode::SPACE, &[], &[Modifier::LEFT_SHIFT]);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.push_key(KeyCode::B);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.push_key(KeyCode::B);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.push_key(KeyCode::N);
-            }
-
-            _ => {}
-        }
-    }
-
-    fn on_release(
-        &mut self,
-        prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-        match (prev.last_perm_layer, prev.layer) {
-            temp_layer!(KeymapLayer::LowerAnsi) => {
-                keyboard_report_state.pop_key(KeyCode::GRAVE);
-            }
-            temp_layer!(KeymapLayer::Lower) => {
-                keyboard_report_state.pop_temp_modifiers();
-                keyboard_report_state.pop_key(KeyCode::EQUALS);
-            }
-            temp_layer!(KeymapLayer::LowerSeMac) => {
-                keyboard_report_state.pop_temp_modifiers();
-                keyboard_report_state.pop_key(KeyCode::SPACE);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.pop_key(KeyCode::B);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.pop_key(KeyCode::B);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.pop_key(KeyCode::N);
-            }
-
-            _ => {}
-        }
-    }
-}
-
-impl KeyboardButton for RightRow3Col0 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, _producer: &Producer) {
-        keyboard_report_state.push_modifier(Modifier::LEFT_CONTROL);
-    }
-
-    fn on_release(
-        &mut self,
-        _prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-        keyboard_report_state.pop_modifier(Modifier::LEFT_CONTROL);
-    }
-}
-
-impl KeyboardButton for RightRow3Col1 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, producer: &Producer) {
-        match (
-            keyboard_report_state.last_perm_layer,
-            keyboard_report_state.active_layer,
-        ) {
-            temp_layer!(KeymapLayer::Settings) => {}
-            base_layer!(KeymapLayer::DvorakSe) => {
-                keyboard_report_state.push_layer_with_fallback(KeymapLayer::Settings);
-                push_layer_change(producer, KeymapLayer::DvorakSe);
-            }
-            base_layer!(KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.push_layer_with_fallback(KeymapLayer::Settings);
-                push_layer_change(producer, KeymapLayer::DvorakSeMac);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.push_layer_with_fallback(KeymapLayer::Settings);
-                push_layer_change(producer, KeymapLayer::DvorakAnsi);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.push_layer_with_fallback(KeymapLayer::Settings);
-                push_layer_change(producer, KeymapLayer::QwertyGaming);
-            }
-
-            _ => {}
-        }
-    }
-
-    fn on_release(
-        &mut self,
-        prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        producer: &Producer,
-    ) {
-        match (prev.last_perm_layer, prev.layer) {
-            temp_layer!(KeymapLayer::Settings) => {
-                keyboard_report_state.pop_layer(KeymapLayer::Settings);
-                push_layer_change(producer, keyboard_report_state.active_layer);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {}
-            base_layer!(KeymapLayer::DvorakAnsi) => {}
-            base_layer!(KeymapLayer::QwertyGaming) => {}
-
-            _ => {}
-        }
-    }
-}
-
-impl KeyboardButton for RightRow3Col2 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, _producer: &Producer) {
-        keyboard_report_state.push_modifier(Modifier::RIGHT_ALT);
-    }
-
-    fn on_release(
-        &mut self,
-        _prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-        keyboard_report_state.pop_modifier(Modifier::RIGHT_ALT);
-    }
-}
-
-impl KeyboardButton for RightRow3Col3 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, producer: &Producer) {
-        match (
-            keyboard_report_state.last_perm_layer,
-            keyboard_report_state.active_layer,
-        ) {
-            temp_layer!(KeymapLayer::Raise) => {}
-            base_layer!(KeymapLayer::DvorakSe) => {
-                keyboard_report_state.push_layer_with_fallback(KeymapLayer::Raise);
-                push_layer_change(producer, KeymapLayer::DvorakSe);
-            }
-            base_layer!(KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.push_layer_with_fallback(KeymapLayer::Raise);
-                push_layer_change(producer, KeymapLayer::DvorakSeMac);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.push_layer_with_fallback(KeymapLayer::Raise);
-                push_layer_change(producer, KeymapLayer::DvorakAnsi);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.push_layer_with_fallback(KeymapLayer::Raise);
-                push_layer_change(producer, KeymapLayer::QwertyGaming);
-            }
-
-            _ => {}
-        }
-    }
-
-    fn on_release(
-        &mut self,
-        prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        producer: &Producer,
-    ) {
-        match (prev.last_perm_layer, prev.layer) {
-            temp_layer!(KeymapLayer::Raise) => {
-                keyboard_report_state.pop_layer(KeymapLayer::Raise);
-                push_layer_change(producer, keyboard_report_state.active_layer);
-            }
-            base_layer!(KeymapLayer::DvorakSe) => {}
-            base_layer!(KeymapLayer::DvorakAnsi) => {}
-            base_layer!(KeymapLayer::QwertyGaming) => {}
-
-            _ => {}
-        }
-    }
-}
-
-impl KeyboardButton for RightRow3Col4 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, producer: &Producer) {
-        match (
-            keyboard_report_state.last_perm_layer,
-            keyboard_report_state.active_layer,
-        ) {
-            temp_layer!(KeymapLayer::Num) => {}
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.push_key(KeyCode::I);
-            }
-            base_layer!(KeymapLayer::DvorakSe) => {
-                keyboard_report_state.push_layer_with_fallback(KeymapLayer::Num);
-                push_layer_change(producer, KeymapLayer::DvorakSe);
-            }
-            base_layer!(KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.push_layer_with_fallback(KeymapLayer::Num);
-                push_layer_change(producer, KeymapLayer::DvorakSeMac);
-            }
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.push_layer_with_fallback(KeymapLayer::Num);
-                push_layer_change(producer, KeymapLayer::DvorakAnsi);
-            }
-
-            _ => {}
-        }
-    }
-
-    fn on_release(
-        &mut self,
-        prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        producer: &Producer,
-    ) {
-        match (prev.last_perm_layer, prev.layer) {
-            temp_layer!(KeymapLayer::Num) => {
-                keyboard_report_state.pop_layer(KeymapLayer::Num);
-                push_layer_change(producer, keyboard_report_state.active_layer);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.push_key(KeyCode::I);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {}
-            base_layer!(KeymapLayer::DvorakAnsi) => {}
-            _ => {}
-        }
-    }
-}
-
-impl KeyboardButton for RightRow3Col5 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, _producer: &Producer) {
-        match (
-            keyboard_report_state.last_perm_layer,
-            keyboard_report_state.active_layer,
-        ) {
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.push_key(KeyCode::SPACE);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.push_key(KeyCode::SPACE);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.push_key(KeyCode::G);
-            }
-            _ => {}
-        }
-    }
-
-    fn on_release(
-        &mut self,
-        prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-        match (prev.last_perm_layer, prev.layer) {
-            base_layer!(KeymapLayer::DvorakAnsi) => {
-                keyboard_report_state.pop_key(KeyCode::SPACE);
-            }
-            base_layer!(KeymapLayer::DvorakSe | KeymapLayer::DvorakSeMac) => {
-                keyboard_report_state.pop_key(KeyCode::SPACE);
-            }
-            base_layer!(KeymapLayer::QwertyGaming) => {
-                keyboard_report_state.pop_key(KeyCode::G);
-            }
-            _ => {}
-        }
-    }
-}
-
-impl KeyboardButton for RightRow4Col1 {
-    fn on_press(&mut self, _keyboard_report_state: &mut KeyboardReportState, _producer: &Producer) {
-    }
-
-    fn on_release(
-        &mut self,
-        _prev: LastPressState,
-        _keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-    }
-    // Rotary encoder is here, no key
-}
-
-impl KeyboardButton for RightRow4Col2 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, _producer: &Producer) {
-        keyboard_report_state.push_key(KeyCode::N2);
-    }
-
-    fn on_release(
-        &mut self,
-        _prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-        keyboard_report_state.pop_key(KeyCode::N2);
-    }
-}
-
-impl KeyboardButton for RightRow4Col3 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, _producer: &Producer) {
-        keyboard_report_state.push_key(KeyCode::N3);
-    }
-
-    fn on_release(
-        &mut self,
-        _prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-        keyboard_report_state.pop_key(KeyCode::N3);
-    }
-}
-
-impl KeyboardButton for RightRow4Col4 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, _producer: &Producer) {
-        keyboard_report_state.push_key(KeyCode::N4);
-    }
-
-    fn on_release(
-        &mut self,
-        _prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-        keyboard_report_state.pop_key(KeyCode::N4);
-    }
-}
-
-impl KeyboardButton for RightRow4Col5 {
-    fn on_press(&mut self, keyboard_report_state: &mut KeyboardReportState, _producer: &Producer) {
-        keyboard_report_state.push_key(KeyCode::N5);
-    }
-
-    fn on_release(
-        &mut self,
-        _prev: LastPressState,
-        keyboard_report_state: &mut KeyboardReportState,
-        _producer: &Producer,
-    ) {
-        keyboard_report_state.pop_key(KeyCode::N5);
-    }
 }
