+++ rust-toolchain.toml
@@ -1,0 +1,4 @@
+[toolchain]
+channel = "stable"
+targets = ["thumbv6m-none-eabi"]
+components = ["rustfmt", "clippy", "rust-src"]
